+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Upgrade to Node 14
+- Migrate to node v14
+- Migrate to sequelize v6
+- Migrate to pg v8
 
 ## [v1.6.0](https://github.com/nathanab/beaconapi/compare/v1.5.0...v1.6.0)
 - Add likes api